requirements.txt
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
requests==2.32.3
soupsieve==2.7
typing_extensions==4.14.0
urllib3==2.0.7
aiogram==3.20.0
httpx==0.24.1
python-dotenv==1.0.0
pika==1.3.1
aio-pika==9.3.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
pytest==7.4.3
pytest-asyncio==0.21.1
